cookbook/tutorials/requirements.txt
phidata
requests
duckduckgo-search
openai
Deprecated
packaging
yfinance